Adjust inheritance hierarchy of type system
 - Adjust project so that non common bits use folders in the project
 - Add DefType as a base type of MetadataType, and define NoMetadataType to derive from that
   - Add new api to MetadataType to get MetadataType base type. (Note, this means that all MetadataTypes must derive from MetadataTypes)
 - Restructure field layout code so that the logic that is associated with a given class is in a seperate c# file that is per class with extra contents.
 - Move MethodImpls logic in EcmaType to seperate file to match factoring of base class.
