+++ src/Common/src/TypeSystem/Common/DefType.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+
+namespace Internal.TypeSystem
+{
+    /// <summary>
+    /// Type that is logically equivalent to a type which is defined by a TypeDef
+    /// record in an ECMA 335 metadata stream.
+    /// </summary>
+    public abstract partial class DefType : TypeDesc
+    {
+    }
+}
