+++ src/Common/src/TypeSystem/Common/FieldLayout.cs
@@ -60,7 +60,7 @@ namespace Internal.TypeSystem
                 (type.IsSequentialLayout || type.IsExplicitLayout) &&
                 (!type.BaseType.IsValueType && !type.BaseType.IsObject))
             {
-                MetadataType baseType = type.BaseType;
+                MetadataType baseType = type.MetadataBaseType;
 
                 if (type.IsSequentialLayout != baseType.IsSequentialLayout ||
                     type.IsExplicitLayout != baseType.IsExplicitLayout)
@@ -353,7 +353,7 @@ namespace Internal.TypeSystem
             _computedLayout.Offsets = offsets;
         }
 
-        private static int ComputeBytesUsedInParentType(MetadataType type)
+        private static int ComputeBytesUsedInParentType(DefType type)
         {
             int cumulativeInstanceFieldPos = 0;
 
@@ -461,290 +461,6 @@ namespace Internal.TypeSystem
         {
             public int Size;
             public int Alignment;
-        }
-    }
-
-    public partial class MetadataType
-    {
-        private class FieldLayoutFlags
-        {
-            public const int HasContainsPointers = 1;
-            public const int ContainsPointers = 2;
-            public const int HasInstanceFieldLayout = 4;
-            public const int HasStaticFieldLayout = 8;
-        }
-
-        private class StaticBlockInfo
-        {
-            public StaticsBlock NonGcStatics;
-            public StaticsBlock GcStatics;
-            public StaticsBlock ThreadStatics;
-        }
-
-        ThreadSafeFlags _fieldLayoutFlags;
-
-        int _instanceFieldSize;
-        int _instanceByteCount;
-        int _instanceFieldAlignment;
-
-        // Information about various static blocks is rare, so we keep it out of line.
-        StaticBlockInfo _staticBlockInfo;
-
-        public bool ContainsPointers
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasContainsPointers))
-                {
-                    ComputeTypeContainsPointers();
-                }
-                return _fieldLayoutFlags.HasFlags(FieldLayoutFlags.ContainsPointers);
-            }
-        }
-
-        public int InstanceFieldSize
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasInstanceFieldLayout))
-                {
-                    ComputeInstanceFieldLayout();
-                }
-                return _instanceFieldSize;
-            }
-        }
-
-        public int InstanceFieldAlignment
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasInstanceFieldLayout))
-                {
-                    ComputeInstanceFieldLayout();
-                }
-                return _instanceFieldAlignment;
-            }
-        }
-
-        public int InstanceByteCount
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasInstanceFieldLayout))
-                {
-                    ComputeInstanceFieldLayout();
-                }
-                return _instanceByteCount;
-            }
-        }
-
-        public int NonGCStaticFieldSize
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasStaticFieldLayout))
-                {
-                    ComputeStaticFieldLayout();
-                }
-                return _staticBlockInfo == null ? 0 : _staticBlockInfo.NonGcStatics.Size;
-            }
-        }
-
-        public int NonGCStaticFieldAlignment
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasStaticFieldLayout))
-                {
-                    ComputeStaticFieldLayout();
-                }
-                return _staticBlockInfo == null ? 0 : _staticBlockInfo.NonGcStatics.LargestAlignment;
-            }
-        }
-
-        public int GCStaticFieldSize
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasStaticFieldLayout))
-                {
-                    ComputeStaticFieldLayout();
-                }
-                return _staticBlockInfo == null ? 0 : _staticBlockInfo.GcStatics.Size;
-            }
-        }
-
-        public int GCStaticFieldAlignment
-        {
-            get
-            {
-                if (!_fieldLayoutFlags.HasFlags(FieldLayoutFlags.HasStaticFieldLayout))
-                {
-                    ComputeStaticFieldLayout();
-                }
-                return _staticBlockInfo == null ? 0 : _staticBlockInfo.GcStatics.LargestAlignment;
-            }
-        }
-
-        internal void ComputeInstanceFieldLayout()
-        {
-            var computedLayout = FieldLayoutAlgorithm.ComputeInstanceFieldLayout(this);
-
-            _instanceFieldSize = computedLayout.FieldSize;
-            _instanceFieldAlignment = computedLayout.FieldAlignment;
-            _instanceByteCount = computedLayout.ByteCount;
-
-            if (computedLayout.Offsets != null)
-            {
-                foreach (var fieldAndOffset in computedLayout.Offsets)
-                {
-                    Debug.Assert(fieldAndOffset.Field.OwningType == this);
-                    fieldAndOffset.Field.InitializeOffset(fieldAndOffset.Offset);
-                }
-            }
-
-            _fieldLayoutFlags.AddFlags(FieldLayoutFlags.HasInstanceFieldLayout);
-        }
-
-        internal void ComputeStaticFieldLayout()
-        {
-            var computedStaticLayout = FieldLayoutAlgorithm.ComputeStaticFieldLayout(this);
-
-            if (computedStaticLayout.Offsets != null)
-            {
-                Debug.Assert(computedStaticLayout.Offsets.Length > 0);
-
-                var staticBlockInfo = new StaticBlockInfo
-                {
-                    NonGcStatics = computedStaticLayout.NonGcStatics,
-                    GcStatics = computedStaticLayout.GcStatics,
-                    ThreadStatics = computedStaticLayout.ThreadStatics
-                };
-                _staticBlockInfo = staticBlockInfo;
-
-                foreach (var fieldAndOffset in computedStaticLayout.Offsets)
-                {
-                    Debug.Assert(fieldAndOffset.Field.OwningType == this);
-                    fieldAndOffset.Field.InitializeOffset(fieldAndOffset.Offset);
-                }
-            }
-
-            _fieldLayoutFlags.AddFlags(FieldLayoutFlags.HasStaticFieldLayout);
-        }
-
-        private void ComputeTypeContainsPointers()
-        {
-            int flagsToAdd = FieldLayoutFlags.HasContainsPointers;
-
-            if (!IsValueType && HasBaseType && BaseType.ContainsPointers)
-            {
-                _fieldLayoutFlags.AddFlags(flagsToAdd | FieldLayoutFlags.ContainsPointers);
-                return;
-            }
-
-            foreach (var field in GetFields())
-            {
-                if (field.IsStatic)
-                    continue;
-
-                TypeDesc fieldType = field.FieldType;
-                if (fieldType.IsValueType)
-                {
-                    if (fieldType.IsPrimitive)
-                        continue;
-
-                    if (((MetadataType)fieldType).ContainsPointers)
-                    {
-                        flagsToAdd |= FieldLayoutFlags.ContainsPointers;
-                        break;
-                    }
-                }
-                else if (fieldType.HasBaseType || fieldType.IsByRef)
-                {
-                    flagsToAdd |= FieldLayoutFlags.ContainsPointers;
-                    break;
-                }
-            }
-
-            _fieldLayoutFlags.AddFlags(flagsToAdd);
-        }
-    }
-
-    public partial class InstantiatedType
-    {
-        public override ClassLayoutMetadata GetClassLayout()
-        {
-            return _typeDef.GetClassLayout();
-        }
-
-        public override bool IsExplicitLayout
-        {
-            get
-            {
-                return _typeDef.IsExplicitLayout;
-            }
-        }
-
-        public override bool IsSequentialLayout
-        {
-            get
-            {
-                return _typeDef.IsSequentialLayout;
-            }
-        }
-
-        public override bool IsBeforeFieldInit
-        {
-            get
-            {
-                return _typeDef.IsBeforeFieldInit;
-            }
-        }
-
-        public override bool IsModuleType { get { return false; } }
-    }
-
-    public partial class FieldDesc
-    {
-        private int _offset = FieldAndOffset.InvalidOffset;
-
-        public int Offset
-        {
-            get
-            {
-                if (_offset == FieldAndOffset.InvalidOffset)
-                {
-                    if (IsStatic)
-                        OwningType.ComputeStaticFieldLayout();
-                    else
-                        OwningType.ComputeInstanceFieldLayout();
-
-                    if (_offset == FieldAndOffset.InvalidOffset)
-                    {
-                        // Must be a field that doesn't participate in layout (literal?)
-                        throw new BadImageFormatException();
-                    }
-                }
-                return _offset;
-            }
-        }
-
-        public bool HasGCStaticBase
-        {
-            get
-            {
-                if (!FieldType.IsValueType)
-                    return true;
-
-                MetadataType fieldType = FieldType as MetadataType;
-                return fieldType != null && fieldType.ContainsPointers;
-            }
-        }
-
-        internal void InitializeOffset(int offset)
-        {
-            Debug.Assert(_offset == FieldAndOffset.InvalidOffset || _offset == offset);
-            _offset = offset;
         }
     }
