+++ src/Common/src/TypeSystem/Common/InstantiatedType.FieldLayout.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+
+namespace Internal.TypeSystem
+{
+    public partial class InstantiatedType
+    {
+    }
+}
