+++ src/Common/src/TypeSystem/Common/TypeSystemHelpers.cs
@@ -32,12 +32,20 @@ namespace Internal.TypeSystem
             return type.Context.GetPointerType(type);
         }
 
-        static public MetadataType GetClosestDefType(this TypeDesc type)
+        static public DefType GetClosestDefType(this TypeDesc type)
+        {
+            if (type is DefType)
+                return (DefType)type;
+            else
+                return type.BaseType;
+        }
+
+        static public MetadataType GetClosestMetadataType(this TypeDesc type)
         {
             if (type is MetadataType)
                 return (MetadataType)type;
             else
-                return type.BaseType;
+                return type.BaseType.GetClosestMetadataType();
         }
 
         static public int GetElementSize(this TypeDesc type)
