+++ src/Common/src/TypeSystem/Ecma/EcmaType.MethodImpls.cs
@@ -1,0 +1,131 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Collections.Generic;
+using System.Reflection;
+using System.Reflection.Metadata;
+using System.Threading;
+using Debug = System.Diagnostics.Debug;
+
+using Internal.TypeSystem;
+
+namespace Internal.TypeSystem.Ecma
+{
+    // This file has implementations of the .MethodImpl.cs logic from its base type.
+
+    public sealed partial class EcmaType : MetadataType
+    {
+        // Virtual function related functionality
+        public override MethodImplRecord[] FindMethodsImplWithMatchingDeclName(string declName)
+        {
+            MetadataReader metadataReader = _module.MetadataReader;
+            var stringComparer = metadataReader.StringComparer;
+            ArrayBuilder<MethodImplRecord> foundRecords = new ArrayBuilder<MethodImplRecord>();
+
+            foreach (var methodImplHandle in _typeDefinition.GetMethodImplementations())
+            {
+                MethodImplementation methodImpl = metadataReader.GetMethodImplementation(methodImplHandle);
+
+                EntityHandle methodDeclCheckHandle = methodImpl.MethodDeclaration;
+                HandleKind methodDeclHandleKind = methodDeclCheckHandle.Kind;
+
+                // We want to check that the method name matches before actually getting the MethodDesc. For MethodSpecifications
+                // we need to dereference that handle to the underlying member reference to look at name matching.
+                if (methodDeclHandleKind == HandleKind.MethodSpecification)
+                {
+                    methodDeclCheckHandle = metadataReader.GetMethodSpecification((MethodSpecificationHandle)methodDeclCheckHandle).Method;
+                    methodDeclHandleKind = methodDeclCheckHandle.Kind;
+                }
+
+                bool foundRecord = false;
+
+                switch (methodDeclHandleKind)
+                {
+                    case HandleKind.MethodDefinition:
+                        if (stringComparer.Equals(metadataReader.GetMethodDefinition((MethodDefinitionHandle)methodDeclCheckHandle).Name, declName))
+                        {
+                            foundRecord = true;
+                        }
+                        break;
+
+                    case HandleKind.MemberReference:
+                        if (stringComparer.Equals(metadataReader.GetMemberReference((MemberReferenceHandle)methodDeclCheckHandle).Name, declName))
+                        {
+                            foundRecord = true;
+                        }
+                        break;
+
+                    default:
+                        Debug.Assert(false, "unexpected methodDeclHandleKind");
+                        break;
+                }
+
+                if (foundRecord)
+                {
+                    MethodImplRecord newRecord = new MethodImplRecord();
+                    newRecord.Decl = (MethodDesc)_module.GetObject(methodImpl.MethodDeclaration);
+                    newRecord.Body = (MethodDesc)_module.GetObject(methodImpl.MethodBody);
+
+                    foundRecords.Add(newRecord);
+                }
+            }
+
+            if (foundRecords.Count != 0)
+                return foundRecords.ToArray();
+
+            return null;
+        }
+
+        protected override MethodImplRecord[] ComputeVirtualMethodImplsForType()
+        {
+            ArrayBuilder<MethodImplRecord> records = new ArrayBuilder<MethodImplRecord>();
+
+            MetadataReader metadataReader = _module.MetadataReader;
+
+            foreach (var methodImplHandle in _typeDefinition.GetMethodImplementations())
+            {
+                MethodImplementation methodImpl = metadataReader.GetMethodImplementation(methodImplHandle);
+
+                EntityHandle methodDeclCheckHandle = methodImpl.MethodDeclaration;
+                HandleKind methodDeclHandleKind = methodDeclCheckHandle.Kind;
+
+                // We want to check that the type is not an interface matches before actually getting the MethodDesc. 
+                // For MethodSpecifications we need to dereference that handle to the underlying member reference to 
+                // look at the owning type.
+                if (methodDeclHandleKind == HandleKind.MethodSpecification)
+                {
+                    methodDeclCheckHandle = metadataReader.GetMethodSpecification((MethodSpecificationHandle)methodDeclCheckHandle).Method;
+                    methodDeclHandleKind = methodDeclCheckHandle.Kind;
+                }
+
+                MetadataType owningType = null;
+                switch (methodDeclHandleKind)
+                {
+                    case HandleKind.MethodDefinition:
+                        owningType = ((MethodDesc)_module.GetObject(methodDeclCheckHandle)).OwningType as MetadataType;
+                        break;
+
+                    case HandleKind.MemberReference:
+                        EntityHandle owningTypeHandle = metadataReader.GetMemberReference((MemberReferenceHandle)methodDeclCheckHandle).Parent;
+                        owningType = _module.GetObject(owningTypeHandle) as MetadataType;
+                        break;
+
+                    default:
+                        Debug.Assert(false, "unexpected methodDeclHandleKind");
+                        break;
+                }
+
+                if (!owningType.IsInterface)
+                {
+                    MethodImplRecord newRecord = new MethodImplRecord();
+                    newRecord.Decl = (MethodDesc)_module.GetObject(methodImpl.MethodDeclaration);
+                    newRecord.Body = (MethodDesc)_module.GetObject(methodImpl.MethodBody);
+                    records.Add(newRecord);
+                }
+            }
+
+            return records.ToArray();
+        }
+    }
+}
