+++ src/Common/src/TypeSystem/Ecma/EcmaType.cs
@@ -12,7 +12,10 @@ using Internal.TypeSystem;
 
 namespace Internal.TypeSystem.Ecma
 {
-    public sealed class EcmaType : MetadataType
+    /// <summary>
+    /// Override of MetadataType that uses actual Ecma335 metadata.
+    /// </summary>
+    public sealed partial class EcmaType : MetadataType
     {
         EcmaModule _module;
         TypeDefinitionHandle _handle;
@@ -125,7 +128,17 @@ namespace Internal.TypeSystem.Ecma
             return type;
         }
 
-        public override MetadataType BaseType
+        public override DefType BaseType
+        {
+            get
+            {
+                if (_baseType == this)
+                    return InitializeBaseType();
+                return _baseType;
+            }
+        }
+
+        public override MetadataType MetadataBaseType
         {
             get
             {
@@ -430,110 +443,6 @@ namespace Internal.TypeSystem.Ecma
             {
                 return Module.GetGlobalModuleType() == this;
             }
-        }
-
-        // Virtual function related functionality
-        public override MethodImplRecord[] FindMethodsImplWithMatchingDeclName(string declName)
-        {
-            MetadataReader metadataReader = _module.MetadataReader;
-            var stringComparer = metadataReader.StringComparer;
-            ArrayBuilder<MethodImplRecord> foundRecords = new ArrayBuilder<MethodImplRecord>();
-
-            foreach (var methodImplHandle in _typeDefinition.GetMethodImplementations())
-            {
-                MethodImplementation methodImpl = metadataReader.GetMethodImplementation(methodImplHandle);
-
-                EntityHandle methodDeclCheckHandle = methodImpl.MethodDeclaration;
-                HandleKind methodDeclHandleKind = methodDeclCheckHandle.Kind;
-
-                // We want to check that the method name matches before actually getting the MethodDesc. For MethodSpecifications
-                // we need to dereference that handle to the underlying member reference to look at name matching.
-                if (methodDeclHandleKind == HandleKind.MethodSpecification)
-                {
-                    methodDeclCheckHandle = metadataReader.GetMethodSpecification((MethodSpecificationHandle)methodDeclCheckHandle).Method;
-                    methodDeclHandleKind = methodDeclCheckHandle.Kind;
-                }
-
-                bool foundRecord = false;
-
-                switch (methodDeclHandleKind)
-                {
-                    case HandleKind.MethodDefinition:
-                        if (stringComparer.Equals(metadataReader.GetMethodDefinition((MethodDefinitionHandle)methodDeclCheckHandle).Name, declName))
-                        {
-                            foundRecord = true;
-                        }
-                        break;
-
-                    case HandleKind.MemberReference:
-                        if (stringComparer.Equals(metadataReader.GetMemberReference((MemberReferenceHandle)methodDeclCheckHandle).Name, declName))
-                        {
-                            foundRecord = true;
-                        }
-                        break;
-                }
-
-                if (foundRecord)
-                {
-                    MethodImplRecord newRecord = new MethodImplRecord();
-                    newRecord.Decl = (MethodDesc)_module.GetObject(methodImpl.MethodDeclaration);
-                    newRecord.Body = (MethodDesc)_module.GetObject(methodImpl.MethodBody);
-
-                    foundRecords.Add(newRecord);
-                }
-            }
-
-            if (foundRecords.Count != 0)
-                return foundRecords.ToArray();
-
-            return null;
-        }
-
-        protected override MethodImplRecord[] ComputeVirtualMethodImplsForType()
-        {
-            ArrayBuilder<MethodImplRecord> records = new ArrayBuilder<MethodImplRecord>();
-
-            MetadataReader metadataReader = _module.MetadataReader;
-
-            foreach (var methodImplHandle in _typeDefinition.GetMethodImplementations())
-            {
-                MethodImplementation methodImpl = metadataReader.GetMethodImplementation(methodImplHandle);
-
-                EntityHandle methodDeclCheckHandle = methodImpl.MethodDeclaration;
-                HandleKind methodDeclHandleKind = methodDeclCheckHandle.Kind;
-
-                // We want to check that the type is not an interface matches before actually getting the MethodDesc. 
-                // For MethodSpecifications we need to dereference that handle to the underlying member reference to 
-                // look at the owning type.
-                if (methodDeclHandleKind == HandleKind.MethodSpecification)
-                {
-                    methodDeclCheckHandle = metadataReader.GetMethodSpecification((MethodSpecificationHandle)methodDeclCheckHandle).Method;
-                    methodDeclHandleKind = methodDeclCheckHandle.Kind;
-                }
-
-                MetadataType owningType = null;
-                switch (methodDeclHandleKind)
-                {
-                    case HandleKind.MethodDefinition:
-                        owningType = ((MethodDesc)_module.GetObject(methodDeclCheckHandle)).OwningType as MetadataType;
-                        break;
-
-                    case HandleKind.MemberReference:
-                        EntityHandle owningTypeHandle = metadataReader.GetMemberReference((MemberReferenceHandle)methodDeclCheckHandle).Parent;
-                        owningType = _module.GetObject(owningTypeHandle) as MetadataType;
-                        break;
-                }
-
-                if (!owningType.IsInterface)
-                {
-                    MethodImplRecord newRecord = new MethodImplRecord();
-                    newRecord.Decl = (MethodDesc)_module.GetObject(methodImpl.MethodDeclaration);
-                    newRecord.Body = (MethodDesc)_module.GetObject(methodImpl.MethodBody);
-                    records.Add(newRecord);
-                }
-            }
-
-            return records.ToArray();
         }
     }
 }
