+++ src/Common/src/TypeSystem/NoMetadata/NoMetadataType.cs
@@ -1,0 +1,22 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Collections.Generic;
+using System.Reflection;
+using System.Reflection.Metadata;
+using System.Threading;
+using Debug = System.Diagnostics.Debug;
+
+using Internal.TypeSystem;
+
+namespace Internal.TypeSystem.NoMetadata
+{
+    /// <summary>
+    /// Base type for types that had metadata at one point, but that metadata is 
+    /// not accessible for the lifetime of the TypeSystemContext
+    /// </summary>
+    public abstract class NoMetadataType: DefType
+    {
+    }
+}
